refactor(worker): extract agent-session resume cluster (A3)
Move _write_approval_record_for_session, _resume_engine_session,
_inject_resume_and_continue, _build_invocation_services,
_repark_continuation, _repark_resumed_yield_outcome to
primer/worker/session_resume_coordinator.py (mirrors
graph_resume_coordinator). pool.py keeps thin delegators; _end_session /
_pause_session and the graph delegators stay. Drop now-unused
yield_runtime/yield_resume_registry/YieldToWorker imports from pool.py.
Pure code motion, no behaviour change.
